Merge v0.8-cs

+++ DLWP/remap/cubesphere.py
@@ -40,16 +40,20 @@ class CubeSphereRemap(_BaseRemap):
     Implement tools for remapping to and from a cubed sphere using TempestRemap executables.
     """
 
-    def __init__(self, path_to_remapper=None):
+    def __init__(self, path_to_remapper=None, to_netcdf4=True, verbose=True):
         """
         Initialize a CubeSphereRemap object.
 
         :param path_to_remapper: str: path to the TempestRemap executables
+        :param to_netcdf4: bool: if True, also use 'ncks' command to convert remapped files to netCDF4
+        :param verbose: bool: print commands and progress
         """
         super(CubeSphereRemap, self).__init__(path_to_remapper=path_to_remapper)
         self.remapper = os.path.join(self.path_to_remapper, 'ApplyOfflineMap')
         self.map = None
         self.inverse_map = None
+        self.to_netcdf4 = to_netcdf4
+        self.verbose = verbose
         self._lat = None
         self._lon = None
         self._res = None
@@ -70,7 +74,8 @@ class CubeSphereRemap(_BaseRemap):
             self.inverse_map = inverse_map_name
             self._inverse_map_exists = True
 
-    def generate_offline_maps(self, lat, lon, res, map_name=None, inverse_map_name=None):
+    def generate_offline_maps(self, lat, lon, res, map_name=None, inverse_map_name=None, inverse_lat=False,
+                              remove_meshes=True):
         """
         Generate offline maps for cubed sphere remapping.
 
@@ -79,6 +84,8 @@ class CubeSphereRemap(_BaseRemap):
         :param res: int: number of points on a side of each cube face
         :param map_name: str: file name of the forward map
         :param inverse_map_name: str: file name of the inverse map
+        :param inverse_lat: if True, then the latitudes in the data file are monotonically decreasing
+        :param remove_meshes: if True, remove the temporary meshes generated while making the offline maps
         :return:
         """
         assert int(lat) > 0
@@ -96,32 +103,55 @@ class CubeSphereRemap(_BaseRemap):
         else:
             self.inverse_map = None
 
-        print('CubeSphereRemap: generating offline forward map...')
+        if self.verbose:
+            print('CubeSphereRemap: generating offline forward map...')
         try:
-            subprocess.check_output([os.path.join(self.path_to_remapper, 'GenerateRLLMesh'),
-                                     '--lat', str(self._lat), '--lon', str(self._lon), '--file', 'outLL.g'])
+            cmd = [os.path.join(self.path_to_remapper, 'GenerateRLLMesh'),
+                   '--lat', str(self._lat), '--lon', str(self._lon), '--file', 'outLL.g']
+            if inverse_lat:
+                cmd = cmd + ['--lat_begin', '90', '--lat_end', '-90']
+            if self.to_netcdf4:
+                cmd = cmd + ['--out_format', 'Netcdf4']
+            if self.verbose:
+                print(' '.join(cmd))
+            subprocess.check_output(cmd)
         except subprocess.CalledProcessError as e:
             print('An error occurred while generating the lat-lon mesh.')
             print(e.output)
             raise
         try:
-            subprocess.check_output([os.path.join(self.path_to_remapper, 'GenerateCSMesh'),
-                                     '--res', str(self._res), '--file', 'outCS.g'])
+            cmd = [os.path.join(self.path_to_remapper, 'GenerateCSMesh'),
+                   '--res', str(self._res), '--file', 'outCS.g']
+            if self.to_netcdf4:
+                cmd = cmd + ['--out_format', 'Netcdf4']
+            if self.verbose:
+                print(' '.join(cmd))
+            subprocess.check_output(cmd)
         except subprocess.CalledProcessError as e:
             print('An error occurred while generating the cube sphere mesh.')
             print(e.output)
             raise
         try:
-            subprocess.check_output([os.path.join(self.path_to_remapper, 'GenerateOverlapMesh'),
-                                     '--a', 'outLL.g', '--b', 'outCS.g', '--out', 'ov_LL_CS.g'])
+            cmd = [os.path.join(self.path_to_remapper, 'GenerateOverlapMesh'),
+                   '--a', 'outLL.g', '--b', 'outCS.g', '--out', 'ov_LL_CS.g']
+            if self.to_netcdf4:
+                cmd = cmd + ['--out_format', 'Netcdf4']
+            if self.verbose:
+                print(' '.join(cmd))
+            subprocess.check_output(cmd)
         except subprocess.CalledProcessError as e:
             print('An error occurred while generating the overlap mesh.')
             print(e.output)
             raise
         try:
-            subprocess.check_output([os.path.join(self.path_to_remapper, 'GenerateOfflineMap'),
-                                     '--in_mesh', 'outLL.g', '--out_mesh', 'outCS.g', '--ov_mesh', 'ov_LL_CS.g',
-                                     '--in_np', '1', '--in_type', 'FV', '--out_type', 'FV', '--out_map', self.map])
+            cmd = [os.path.join(self.path_to_remapper, 'GenerateOfflineMap'),
+                   '--in_mesh', 'outLL.g', '--out_mesh', 'outCS.g', '--ov_mesh', 'ov_LL_CS.g',
+                   '--in_np', '1', '--in_type', 'FV', '--out_type', 'FV', '--out_map', self.map]
+            if self.to_netcdf4:
+                cmd = cmd + ['--out_format', 'Netcdf4']
+            if self.verbose:
+                print(' '.join(cmd))
+            subprocess.check_output(cmd)
         except subprocess.CalledProcessError as e:
             print('An error occurred while generating the offline map.')
             print(e.output)
@@ -130,27 +160,38 @@ class CubeSphereRemap(_BaseRemap):
 
         print('CubeSphereRemap: generating offline inverse map...')
         try:
-            subprocess.check_output([os.path.join(self.path_to_remapper, 'GenerateOverlapMesh'),
-                                     '--a', 'outCS.g', '--b', 'outLL.g', '--out', 'ov_CS_LL.g'])
+            cmd = [os.path.join(self.path_to_remapper, 'GenerateOverlapMesh'),
+                   '--a', 'outCS.g', '--b', 'outLL.g', '--out', 'ov_CS_LL.g']
+            if self.to_netcdf4:
+                cmd = cmd + ['--out_format', 'Netcdf4']
+            if self.verbose:
+                print(' '.join(cmd))
+            subprocess.check_output(cmd)
         except subprocess.CalledProcessError as e:
             print('An error occurred while generating the overlap mesh.')
             print(e.output)
             raise
         try:
-            subprocess.check_output([os.path.join(self.path_to_remapper, 'GenerateOfflineMap'),
-                                     '--in_mesh', 'outCS.g', '--out_mesh', 'outLL.g', '--ov_mesh', 'ov_CS_LL.g',
-                                     '--in_np', '1', '--in_type', 'FV', '--out_type', 'FV', '--out_map',
-                                     self.inverse_map])
+            cmd = [os.path.join(self.path_to_remapper, 'GenerateOfflineMap'),
+                   '--in_mesh', 'outCS.g', '--out_mesh', 'outLL.g', '--ov_mesh', 'ov_CS_LL.g',
+                   '--in_np', '1', '--in_type', 'FV', '--out_type', 'FV', '--out_map', self.inverse_map]
+            if self.to_netcdf4:
+                cmd = cmd + ['--out_format', 'Netcdf4']
+            if self.verbose:
+                print(' '.join(cmd))
+            subprocess.check_output(cmd)
         except subprocess.CalledProcessError as e:
             print('An error occurred while generating the offline map.')
             print(e.output)
             raise
 
-        for f in ['outLL.g', 'outCS.g', 'ov_LL_CS.g', 'ov_CS_LL.g']:
-            os.remove(f)
+        if remove_meshes:
+            for f in ['outLL.g', 'outCS.g', 'ov_LL_CS.g', 'ov_CS_LL.g']:
+                os.remove(f)
 
         self._inverse_map_exists = True
-        print('CubeSphereRemap: succesfully generated offline maps (%s, %s)' % (self.map, self.inverse_map))
+        if self.verbose:
+            print('CubeSphereRemap: succesfully generated offline maps (%s, %s)' % (self.map, self.inverse_map))
 
     def remap(self, input_file, output_file, *args):
         """
@@ -165,17 +206,22 @@ class CubeSphereRemap(_BaseRemap):
                              "'assign_maps' functions first")
         elif not(os.path.exists(self.map)):
             raise FileNotFoundError(self.map)
+        if self.verbose:
+            print('CubeSphereRemap: applying forward map...')
 
         try:
-            subprocess.check_output([os.path.join(self.path_to_remapper, 'ApplyOfflineMap'),
-                                     '--in_data', input_file, '--out_data', output_file, '--map', self.map]
-                                    + list(args))
+            cmd = [os.path.join(self.path_to_remapper, 'ApplyOfflineMap'),
+                   '--in_data', input_file, '--out_data', output_file, '--map', self.map] + list(args)
+            if self.verbose:
+                print(' '.join(cmd))
+            subprocess.check_output(cmd)
         except subprocess.CalledProcessError as e:
             print('An error occurred while applying the offline map.')
             print(e.output)
             raise
 
-        print('CubeSphereRemap: successfully remapped data into %s' % output_file)
+        if self.verbose:
+            print('CubeSphereRemap: successfully remapped data into %s' % output_file)
 
     def inverse_remap(self, input_file, output_file, *args):
         """
@@ -190,20 +236,24 @@ class CubeSphereRemap(_BaseRemap):
                              "'assign_maps' functions first")
         elif not (os.path.exists(self.inverse_map)):
             raise FileNotFoundError(self.inverse_map)
+        if self.verbose:
+            print('CubeSphereRemap: applying inverse map...')
 
         try:
-            subprocess.check_output([os.path.join(self.path_to_remapper, 'ApplyOfflineMap'),
-                                     '--in_data', input_file, '--out_data', output_file, '--map', self.inverse_map]
-                                    + list(args))
+            cmd = [os.path.join(self.path_to_remapper, 'ApplyOfflineMap'),
+                   '--in_data', input_file, '--out_data', output_file, '--map', self.inverse_map] + list(args)
+            if self.verbose:
+                print(' '.join(cmd))
+            subprocess.check_output(cmd)
         except subprocess.CalledProcessError as e:
             print('An error occurred while applying the offline map.')
             print(e.output)
             raise
 
-        print('CubeSphereRemap: successfully inverse remapped data into %s' % output_file)
+        if self.verbose:
+            print('CubeSphereRemap: successfully inverse remapped data into %s' % output_file)
 
-    @staticmethod
-    def convert_to_faces(input_file, output_file, coord_file=None, chunking=None):
+    def convert_to_faces(self, input_file, output_file, coord_file=None, chunking=None):
         """
         Convert a file in cubed-sphere coordinates to contain dimensions for the face number and height/width of each
         of the six cube faces. Very useful for applying convolutions on the faces.
@@ -217,6 +267,9 @@ class CubeSphereRemap(_BaseRemap):
         """
         # Open the dataset to convert
         ds = xr.open_dataset(input_file)
+        if self.verbose:
+            print('CubeSphereRemap.convert_to_faces: loading data to memory...')
+        ds.load()
 
         # First, assign any coordinates missing from the input file from the coordinate file, if provided.
         if coord_file is not None:
@@ -235,20 +288,22 @@ class CubeSphereRemap(_BaseRemap):
 
         # Assign the new coordinate and transpose
         new_dims = tuple([d for d in ds.dims.keys() if d != 'ncol']) + ('height', 'width', 'face')
-        print('CubeSphereRemap: assigning new coordinates to dataset')
+        if self.verbose:
+            print('CubeSphereRemap.convert_to_faces: assigning new coordinates to dataset')
         ds_new = ds.assign_coords(ncol=face_index).unstack('ncol').transpose(*new_dims)
 
         # Export to a new file
-        print('CubeSphereRemap: exporting data to file %s' % output_file)
+        if self.verbose:
+            print('CubeSphereRemap.convert_to_faces: exporting data to file %s...' % output_file)
         if chunking is not None:
             ds_new = to_chunked_dataset(ds_new, chunking)
         ds_new.to_netcdf(output_file)
 
-        print('CubeSphereRemap: succesfully exported reformatted data')
+        if self.verbose:
+            print('CubeSphereRemap.convert_to_faces: successfully exported reformatted data')
         return ds_new
 
-    @staticmethod
-    def convert_from_faces(input_file, output_file, chunking=None):
+    def convert_from_faces(self, input_file, output_file, chunking=None):
         """
         Revert a file containing height, width, face dimensions into the default 'ncol' dimension so that it can be
         inverse remapped by the remapper.
@@ -260,6 +315,9 @@ class CubeSphereRemap(_BaseRemap):
         """
         # Open the dataset to convert
         ds = xr.open_dataset(input_file)
+        if self.verbose:
+            print('CubeSphereRemap.convert_from_faces: loading data to memory...')
+        ds.load()
 
         # Transpose the face dimension and stack the face, height, width
         fhw = ('face', 'height', 'width')
@@ -268,10 +326,12 @@ class CubeSphereRemap(_BaseRemap):
         ds_new = ds.transpose(*new_dims).stack(ncol=fhw).reset_index('ncol')
 
         # Export to new file
-        print('CubeSphereRemap: exporting data to file %s' % output_file)
+        if self.verbose:
+            print('CubeSphereRemap.convert_from_faces: exporting data to file %s...' % output_file)
         if chunking is not None:
             ds_new = to_chunked_dataset(ds_new, chunking)
         ds_new.to_netcdf(output_file)
 
-        print('CubeSphereRemap: succesfully exported reformatted data')
+        if self.verbose:
+            print('CubeSphereRemap.convert_from_faces: successfully exported reformatted data')
         return ds_new
